requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
motor==3.6.0
pymongo==4.9.2
PyJWT==2.10.1
argon2-cffi==23.1.0
python-dotenv==1.0.1
cloudinary==1.41.0
python-multipart==0.0.20
